vinyl service + DTO relations

+++ src/main/java/com/themusicshop/app/persistence/repository/VinylRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VinylRepository extends JpaRepository<Vinyl, Integer> {
+public interface VinylRepository extends JpaRepository<Vinyl, Long> {
+
 }
